+++ src/react/dropdowns.jsx
@@ -92,18 +92,18 @@ var Dropdown = React.createClass({
       </div>
     );
   },
-  _click: function (ev) {
+  _click: function(ev) {
     // only left clicks
     if (ev.button !== 0) return;
 
     // exit if toggle button is disabled
     if (this.props.disabled) return;
 
-    setTimeout(function () {
+    setTimeout(function() {
       if (!ev.defaultPrevented) this._toggle();
     }.bind(this), 0);
   },
-  _toggle: function () {
+  _toggle: function() {
     // exit if no menu element
     if (!this.props.children) {
       return util.raiseError('Dropdown menu element not found');
@@ -112,7 +112,7 @@ var Dropdown = React.createClass({
     if (this.state.opened) this._close();
     else this._open();
   },
-  _open: function () {
+  _open: function() {
     // position menu element below toggle button
     var wrapperRect = React.findDOMNode(this).getBoundingClientRect(),
         toggleRect;
@@ -124,16 +124,16 @@ var Dropdown = React.createClass({
       opened: true
     });
   },
-  _close: function () {
+  _close: function() {
     this.setState({
       opened: false
     });
   },
-  _select: function (ev) {
+  _select: function(ev) {
     if (this.props.onClick) this.props.onClick(this, ev);
   },
-  _outsideClick: function (ev) {
-    var isClickInside = React.findDOMNode(this).contains(event.target);
+  _outsideClick: function(ev) {
+    var isClickInside = React.findDOMNode(this).contains(ev.target);
 
     if (!isClickInside) {
       this._close();
@@ -147,7 +147,7 @@ var Dropdown = React.createClass({
  * @class
  */
 var DropdownItem = React.createClass({
-  render: function () {
+  render: function() {
     return (
       <li>
         <a href={ this.props.link || '#' } onClick={ this._click }>
@@ -156,7 +156,7 @@ var DropdownItem = React.createClass({
       </li>
     );
   },
-  _click: function (ev) {
+  _click: function(ev) {
     if (this.props.onClick) this.props.onClick(this, ev);
   }
 });